P. Set oil analysis
In this problem set, you are ask to study the relationship between oil prices and stock prices.
Intuitively, if oil prices decline because of an increase in supply, the economy should benet and
stock prices should go up. On the contrary, if oil prices decline because demand shrinks, then
this is a signal of weak aggregate demand and stock prices should go down. You can nd data
for this problem set on the e-learning page. Note that the data have been downloaded from
Thomson Reuters Datastream, at daily frequency, for the period 12/31/2006 to 9/20/2017. Refer
to Datastream for details regarding the data and to the Excel sheet readme for the name of the
variables.